Add type annotations, fix imports, and improve code quality across codebase
- Add comprehensive type hints to all function signatures and return types
- Configure mypy via pyproject.toml for static type checking
- Convert bare imports to relative imports for proper package structure
- Add closure support and return values to all custom optimizer step() methods
- Fix potential division-by-zero in test functions (max(batch_idx, 1))
- Add null-image guard in read_img.py and re-raise exceptions
- Wrap numpy scalar returns with float() for correct return types
- Remove dead code, unused config, and empty __init__.py
- Rename resnet.py to Resnet.py for consistent naming
- Add .mypy_cache/ to .gitignore

+++ common_alg/distance/euclidean.py
@@ -10,5 +10,5 @@ def euclidean(d1: np.ndarray, d2: np.ndarray) -> float:
     returns:
         The Euclidean distance as a float.
     '''
-    return np.sqrt(np.sum((d1-d2)**2))
+    return float(np.sqrt(np.sum((d1-d2)**2)))
